Renaming perform_env.*
