+++ Source/Global/NetworkState.h
